feat: Creating controller from create a new comment feature
