+++ src/repositories/CreateCommentService/CreateCommentsService.ts
@@ -2,7 +2,7 @@ import { Comment } from '@prisma/client'
 import { ISaveFunctionGeneric, PrismaClientInstance } from '../GenericsInterfaces/GenericRepository'
 
 export class CreateCommentService implements ISaveFunctionGeneric<Comment>{
-	async save(data: any): Promise<any> {
+	async save(data: any): Promise<Comment> {
 		await PrismaClientInstance.comment.create({
 			data: data
 		})
